requirements.txt
python-telegram-bot==20.7
yt_dlp
requests
python-dotenv
